+++ src/main/java/org/eclipse/keyple/card/calypso/crypto/legacysam/LegacySamExtensionService.java
@@ -70,21 +70,26 @@ public final class LegacySamExtensionService implements KeypleCardExtension {
   /**
    * Returns a {@link CardResourceProfileExtension} to be used with the card resource service.
    *
-   * @param legacySamSelection The legacy SAM selection extension to use.
+   * @param legacySamSelectionExtension The legacy SAM selection extension to use.
    * @return A not null reference.
    * @throws IllegalArgumentException If no SAM selection is provided.
    * @since 0.2.0
    */
   public CardResourceProfileExtension createLegacySamResourceProfileExtension(
-      LegacySamSelectionExtension legacySamSelection) {
-    Assert.getInstance().notNull(legacySamSelection, "Legacy SAM selection");
-    return new LegacySamResourceProfileExtensionAdapter(legacySamSelection, null);
+      LegacySamSelectionExtension legacySamSelectionExtension) {
+    Assert.getInstance().notNull(legacySamSelectionExtension, "Legacy SAM selection");
+    if (!(legacySamSelectionExtension instanceof LegacySamSelectionExtensionAdapter)) {
+      throw new IllegalArgumentException(
+          "The provided 'legacySamSelectionExtension' must be an instance of 'LegacySamSelectionExtensionAdapter'");
+    }
+    return new LegacySamResourceProfileExtensionAdapter(
+        (LegacySamSelectionExtensionAdapter) legacySamSelectionExtension, null);
   }
 
   /**
    * Returns a {@link CardResourceProfileExtension} to be used with the card resource service.
    *
-   * @param legacySamSelection The legacy SAM selection extension to use.
+   * @param legacySamSelectionExtension The legacy SAM selection extension to use.
    * @param powerOnDataRegex A regular expression.
    * @return A not null reference.
    * @throws IllegalArgumentException If no SAM selection is provided.
@@ -92,11 +97,16 @@ public final class LegacySamExtensionService implements KeypleCardExtension {
    * @since 0.2.0
    */
   public CardResourceProfileExtension createLegacySamResourceProfileExtension(
-      LegacySamSelectionExtension legacySamSelection, String powerOnDataRegex) {
+      LegacySamSelectionExtension legacySamSelectionExtension, String powerOnDataRegex) {
     Assert.getInstance()
-        .notNull(legacySamSelection, "Legacy SAM selection")
+        .notNull(legacySamSelectionExtension, "Legacy SAM selection")
         .notEmpty(powerOnDataRegex, "powerOnDataRegex");
-    return new LegacySamResourceProfileExtensionAdapter(legacySamSelection, powerOnDataRegex);
+    if (!(legacySamSelectionExtension instanceof LegacySamSelectionExtensionAdapter)) {
+      throw new IllegalArgumentException(
+          "The provided 'legacySamSelectionExtension' must be an instance of 'LegacySamSelectionExtensionAdapter'");
+    }
+    return new LegacySamResourceProfileExtensionAdapter(
+        (LegacySamSelectionExtensionAdapter) legacySamSelectionExtension, powerOnDataRegex);
   }
 
   /**